# Add an option to control the size of sendQ buffer