In calltypes using `NodeStream`, the size of the buffer in the `sendQ` channel can have a major impact on performance for some calltypes. In particular, a small buffer yields lower latency, while a larger buffer yields higher throughput.

This PR adds a new `ManagerOption`, `WithSendBufferSize`, that gives control over the size of the buffer on this channel.

I have done some benchmarks comparing the new implementation of Multicast (#65) with the old version, using different buffer sizes. The benchmarks were performed on some of the `pitter` machines at UiS, with 1 client and 4 servers.

![throughput](https://user-images.githubusercontent.com/10762082/84387939-0d326d80-abf4-11ea-9eb7-110bf8264e6f.png)
![latency](https://user-images.githubusercontent.com/10762082/84387950-115e8b00-abf4-11ea-9aa1-53ba6ef8ad57.png)

Raw data:

Baseline (old multicast):

```
Benchmark    Througput           Latency    Std.dev    Client+Servers                    
Multicast    86596.45 ops/sec    0.44 ms    0.82 ms    3101 B/op         61 allocs/op 
```

New Multicast:

```
Buffer size: 0
Benchmark    Througput           Latency    Std.dev    Client+Servers                    
Multicast    55255.25 ops/sec    0.27 ms    0.51 ms    4399 B/op         83 allocs/op    

Buffer size: 1
Benchmark    Througput           Latency    Std.dev    Client+Servers                    
Multicast    62638.19 ops/sec    0.31 ms    0.54 ms    4464 B/op         83 allocs/op    

Buffer size: 10
Benchmark    Througput            Latency    Std.dev    Client+Servers                    
Multicast    100948.09 ops/sec    0.50 ms    0.89 ms    4529 B/op         83 allocs/op    

Buffer size: 100
Benchmark    Througput            Latency    Std.dev    Client+Servers                    
Multicast    156563.76 ops/sec    1.67 ms    1.85 ms    4547 B/op         83 allocs/op    

Buffer size: 1000
Benchmark    Througput            Latency    Std.dev     Client+Servers                    
Multicast    153235.60 ops/sec    9.49 ms    17.98 ms    4652 B/op         83 allocs/op    
```