@Raytar 2020-06-11
Also note that I had to modify the "old" multicast in order to make it work properly. In particular, I removed the extra goroutine that is started per node for each multicast call:

```go
func (c *Configuration) {{$method}}(in *{{$in}}) error {
	for _, node := range c.nodes {
		err := node.{{unexport $method}}Client.Send(in)
		if err == nil {
			continue
		}
		if c.mgr.logger != nil {
			c.mgr.logger.Printf("%d: {{$method}} stream send error: %v", node.id, err)
		}
	}
	return nil
}
```
Without this tweak, it becomes impossible to know whether a multicast has been sent. Thus, if executing a multicast call in a loop, the system can run out of memory because it creates new goroutines to send messages faster than messages are sent.